fix: handle non-UTF-8 logcat output and u8 overflow
- Fix u8 overflow in millisecond parsing (src/logcat.rs:199):
  Promote to u32 before *100 multiplication to avoid overflow
  when ms >= 300.

- Replace BufReader::lines() with read_until + from_utf8_lossy
  to handle binary data in Android logcat messages without
  crashing the reader task.

- Log rejected lines and reader exit reason to tlog-rejected.log
  instead of stderr to avoid TUI corruption.

+++ src/logcat.rs
@@ -196,7 +196,7 @@ pub fn parse_line(line: &str) -> Option<LogEntry> {
     if bytes.len() < 18 || !bytes[15].is_ascii_digit() || !bytes[16].is_ascii_digit() || !bytes[17].is_ascii_digit() {
         return None;
     }
-    let ms: u32 = ((bytes[15] - b'0') * 100 + (bytes[16] - b'0') * 10 + (bytes[17] - b'0')) as u32;
+    let ms: u32 = (bytes[15] - b'0') as u32 * 100 + (bytes[16] - b'0') as u32 * 10 + (bytes[17] - b'0') as u32;
 
     // Check space at pos 18
     if bytes.get(18)? != &b' ' {
